Add no scrollbar variant
There is currently no way to disable scroll bars in fltk-rs. An enum in Rust is
an actual new type wheres in Cpp it is just an alias for an int. This allows to
use `0` as scrollbar type (which disables it) in fltk even though it is not part
of the enum.

In Rust this is not possible. Hence, we add an explicit enum variant for
disabling scroll bars with value `0`.