Use (de)activate to show file is modified or not.
Variable saved is renamed as modified, to indicate that the content of the buffer is changed.
Menu item File/Quit is now colored red after the text buffer is modified and File/Save is activated, as it was initially started deactivated. Will be deactivated after a succesful File/save action.
Furthermore, the filename is stored as an option of a path, which forces a Rust-style handling with match or unwrap.
Solves the bug that quitting a modified buffer and then cancelling the file save dialog result in loss of work.