editor, editor2 examples: Add accelerators
The editor examples were not taking full advantage of accelerators:
menus had accelerators, but their items didn't, making them much less
useful.

They're easy enough to add, and I think it's important that new
developers see them in the examples and know that they work and how to
use them, so that they also use them... because if there is something I
could use less of, is GUI applications being needlessly awkward to use
with a keyboard.